requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# LangChain & AI
langchain>=0.1.6
langchain-google-genai>=0.0.6
langchain-community>=0.0.18
google-generativeai>=0.3.2
sentence-transformers>=2.2.0

# Vector Store
chromadb>=0.4.22

# Document Processing
PyMuPDF>=1.24.0
python-docx
openpyxl
xlrd

# URL Content Fetching
requests>=2.31.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0

GitPython
python-docx

# MCP Server
mcp>=1.0.0
